linode_vpn.tf, scripts/*: fixed race conditions, improved organization
fixed race condition:

when destroying an instance the destroy provisioner is called and
updates the ssh config file located in the login folder. Needed
only one instance to modify this file at a time so i created a
lockfile script and polled until the file wasn't being used by
a resource.

Signed-off-by: Thomas Turner <thomastdt@gmail.com>